# Speedup patch embedding by moving coord creation on device