While experimenting with image segmentation and using the `prun` function on `predictor.set_image`, I discovered a significant performance issue in the `get_rel_pos` function.

The problem arises from the lack of device assignment for the tensor containing relative positions, causing unnecessary data transfers when using accelerators like GPUs.

The solution is rather simple: create the coordinate tensor on the same device as the underlying data. 
In my specific case (vit-b, P100 GPU, 128x128 image size), this change reduced the `get_rel_pos` execution time from 197ms to 3ms, cutting down the `predictor.set_image` runtime from 220ms to 20ms.

Given that there's no reason to keep the coordinate tensor creation outside the device, I suggest incorporating this change into the official release for the benefit of all users.